pyproject.toml
[project]
name = "sqs-consumer"
version = "0.1.0"
description = "Sqs message consumer and publish to rabbitMQ"
author ="Abilash Suresh"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aioboto3>=15.5.0",
    "pika>=1.3.2",
    "python-dotenv>=1.2.2",
]
